Move cluster code to separate file
